Rows validation, Double -> Float, code cleanup
- Added validation on AddTankVC (using extension) - changes text color depending on row's Rules
- Changed from Double to Float (Double precision is not needed there)
- Cleaned a bit of code

+++ iAquarium2/iAquarium2/DataModels/NSManagedObjects/WaterParameter+CoreDataProperties.swift
@@ -17,13 +17,13 @@ extension WaterParameter {
         return NSFetchRequest<WaterParameter>(entityName: "WaterParameter")
     }
 
-    @NSManaged public var cl2Value: Double
-    @NSManaged public var ghValue: Double
-    @NSManaged public var khValue: Double
-    @NSManaged public var no2Value: Double
-    @NSManaged public var no3Value: Double
-    @NSManaged public var phValue: Double
-    @NSManaged public var temp: Double
+    @NSManaged public var cl2Value: Float
+    @NSManaged public var ghValue: Float
+    @NSManaged public var khValue: Float
+    @NSManaged public var no2Value: Float
+    @NSManaged public var no3Value: Float
+    @NSManaged public var phValue: Float
+    @NSManaged public var temp: Float
     @NSManaged public var ofMeasurement: Measurement?
 
 }
